requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.1.0
transformers>=4.40.0
datasets>=2.14.0
accelerate>=1.0.0  # Ensure compatibility with latest transformers

# Data processing
tqdm>=4.65.0
requests>=2.31.0
aiohttp>=3.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# LLM Fine-tuning with Unsloth
unsloth>=2024.6
peft>=0.8.0
bitsandbytes>=0.42.0
trl>=0.7.0

# Inference with vLLM
vllm>=0.5.0

# Configuration
pyyaml>=6.0

# Logging
wandb>=0.16.0  # Optional: for experiment tracking

# Utilities
pydantic>=2.0.0
typing-extensions>=4.8.0
